+++ styles/Layout.module.scss
@@ -20,7 +20,7 @@
   min-height: 100vh;
   overflow: auto;
   padding: 2rem;
-  padding-bottom: 6rem;
+  padding-bottom: 4rem;
 }
 
 .header_main {
